BufferValue::SizeFunction should be AnyInvocable
SizeFunction is often passed around with std::move, but rarely if ever can it
actally be 'moved'. We can't use move captures with the old definition of
SizeFunction and therefore have to use clever workarounds for lifetime issues
with reference captures. This patch aims to change this once and for all. While
the size function can still be passed around by reference/pointer, we explicitly
forbid copies to allow real moves.

PiperOrigin-RevId: 808264834